[APP] Use date in local time for birthdate without browser offset
