+++ mon-pix/app/components/user-certifications-detail-header.js
@@ -2,17 +2,17 @@ import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
+import parseISODateOnly from '../utils/parse-iso-date-only';
 
 export default class UserCertificationsDetailHeader extends Component {
   @service intl;
-  @service intlDateFormatter;
   @service fileSaver;
   @service session;
 
   @tracked tooltipText = this.intl.t('pages.certificate.verification-code.copy');
 
-  get birthdate() {
-    return this.intlDateFormatter.formatDateStringToISO(this.args.certification.birthdate);
+  get birthdateMidnightLocalTime() {
+    return parseISODateOnly(this.args.certification.birthdate);
   }
 
   @action
